frontend/requirements.txt
streamlit
requests
Pillow
pandas